+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Support for local-only contact receivers.
 - [NDMF] `VQT Mesh Flipper` component to flip the mesh of the avatar.
 - [NDMF] `[NDMF] Manual Bake with Android Settings` menu to the right-click menu of the avatar.
 - [NDMF] Feature to select an avatar to upload in the `VQT Avatar Builder`.